// src/lib/api/gen/types/AuthUserDto.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

export type AuthUserDto = {
  /**
   * @type string
   */
  id: string;
  /**
   * @type string
   */
  name: string;
  /**
   * @type string
   */
  email: string;
  /**
   * @type array
   */
  roles: string[];
  /**
   * @type array
   */
  permissions: string[];
};